Bruk JobRunner for distribusjon schedule Og litt rename.
